requirements.txt
numpy==1.19.5
scikit-learn==0.24.2
tensorflow-gpu==2.5.0